Create LabeledControls folder and separate components
